Factor out code that creates Neon events
