+++ studio/src/pages/RequestorPage/LogsTable/LogsTable.tsx
@@ -1,16 +1,9 @@
 import { useOrphanLogs } from "@/hooks";
 import { useOtelTrace } from "@/queries";
-import {
-  getNeonSqlQuery,
-  getResponseBody,
-  isJson,
-  isNeonFetch,
-  safeParseJson,
-} from "@/utils";
-import { useMemo } from "react";
 import { LogsEmptyState } from "./Empty";
 import { LogsTableRow } from "./LogsTableRow";
-import type { LogEntry, NeonEvent } from "./types";
+import { useLogsWithEvents } from "./data";
+import type { LogEntry } from "./types";
 
 type Props = {
   traceId?: string;
@@ -30,37 +23,11 @@ const LogsTableWithTraceId = ({ traceId }: { traceId: string }) => {
   const { data: spans } = useOtelTrace(traceId);
   const logs = useOrphanLogs(traceId, spans ?? []);
 
-  // Here we can insert relevant events that happend, in order to link back to the timeline.
-  // For now, we're just looking for Neon database queries
-  const logsWithEvents = useMemo<LogEntry[]>(() => {
-    const neonSpans = spans?.filter((span) => isNeonFetch(span));
-    const neonEvents: NeonEvent[] =
-      neonSpans?.map((span) => {
-        const responseBody = getResponseBody(span);
-        const rowCount =
-          responseBody && isJson(responseBody)
-            ? Number.parseInt(safeParseJson(responseBody)?.rowCount ?? "") ??
-              null
-            : null;
-        return {
-          id: span.span_id,
-          type: "neon-event",
-          timestamp: span.end_time,
-          sql: getNeonSqlQuery(span),
-          rowCount,
-          duration: span.end_time.getTime() - span.start_time.getTime(),
-        };
-      }) ?? [];
-
-    if (neonEvents?.length) {
-      const result = [...logs, ...neonEvents];
-      return result.sort(
-        (a, b) => a.timestamp.getTime() - b.timestamp.getTime(),
-      );
-    }
-
-    return logs;
-  }, [logs, spans]);
+  // Here we insert relevant events that happened.
+  // For now, we're just looking for Neon database queries.
+  // Jacco is going to add exceptions, then we should consider additional things like
+  // fetches, etc.
+  const logsWithEvents = useLogsWithEvents(spans ?? [], logs);
 
   return <LogsTableContent logs={logsWithEvents} />;
 };
